add images management for every schema

+++ server/routes/images.js
@@ -4,6 +4,10 @@ const path = require("path");
 const fs = require("fs");
 const User = require("../models/user.model");
 const Pet = require("../models/pet.model");
+const Product = require("../models/product.model");
+const Post = require("../models/posts.model");
+const Service = require("../models/services.model");
+const News = require("../models/news.model");
 
 /**
  * Storage delle immagini con regole sul come salvarle
@@ -21,7 +25,7 @@ const storage = multer.diskStorage({
  * Router per la gestione delle immagini profilo degli utenti
  */
 router
-  .route("/avatar")
+  .route("/user")
   /**
    * GET
    * Ritorna l'immagine profilo di un utente
@@ -54,31 +58,53 @@ router
   .post(async (req, res) => {
     try {
       if (req.user != null) {
-        const upload = multer({ storage: storage }).single("image");
-        upload(req, res, async function (err) {
-          if (err) {
-            console.log(err);
-          } else {
-            await User.findOne({ username: req.user.username }).then(
-              async (user) => {
-                await fs.unlink(
-                  __foldername + "/server/Images/" + user.profilePicture,
-                  (err) => {
-                    if (err) console.log(err);
-                  }
-                );
-              }
-            );
-            await User.findOneAndUpdate(
-              { username: req.user.username },
-              {
-                profilePicture: res.req.file.filename,
-              }
-            ).then(() => {
-              res.status(200).send("avatar uploaded correctly");
-            });
-          }
-        });
+        if (req.user.profilePicture != "") {
+          const upload = multer({ storage: storage }).single("image");
+          upload(req, res, async function (err) {
+            if (err) {
+              console.log(err);
+            } else {
+              await User.findOne({ username: req.user.username }).then(
+                async (user) => {
+                  await fs.unlink(
+                    __foldername + "/server/Images/" + user.profilePicture,
+                    (err) => {
+                      if (err) console.log(err);
+                    }
+                  );
+                }
+              );
+              await User.updateOne(
+                { username: req.user.username },
+                {
+                  profilePicture: res.req.file.filename,
+                }
+              ).then(() => {
+                res.status(200).send("avatar uploaded correctly");
+              });
+            }
+          });
+        } else {
+          const upload = multer({ storage: storage }).single("image");
+          let file = res.req.files;
+          console.log(file);
+          upload(req, res, async function (err) {
+            if (err) {
+              console.log(err);
+            } else {
+              // console.log(user);
+              console.log(file);
+              await User.updateOne(
+                { username: req.user.username },
+                {
+                  profilePicture: res.req.file.filename,
+                }
+              ).then(() => {
+                res.status(200).send("avatar uploaded correctly");
+              });
+            }
+          });
+        }
       } else {
         res.status(401).send("unauthorized");
       }
@@ -141,7 +167,11 @@ router
   });
 
 router
-  .route("/petPhotos")
+  .route("/pets")
+  /**
+   * GET
+   * Get all photos of a pet
+   */
   .get(async (req, res) => {
     try {
       if (req.user != null) {
@@ -152,8 +182,10 @@ router
             }
           });
         } else {
-          await Pet.findOne({ _id: req.query.id }).then((pet) => {
-            if (pet.owner == req.user.username) {
+          await Pet.findOne({
+            $and: [{ _id: req.query.id }, { owner: req.user.username }],
+          }).then((pet) => {
+            if (pet) {
               res.status(200).send(pet.pictures);
             } else {
               res.sendStatus(401);
@@ -167,6 +199,10 @@ router
       console.log(error);
     }
   })
+  /**
+   * POST
+   * create new pet with photos
+   */
   .post(async (req, res) => {
     try {
       if (req.user != null) {
@@ -176,8 +212,9 @@ router
             console.log(err);
           } else {
             const imagesAddr = [];
+            console.log(res.req.files);
             for (let i = 0; i < res.req.files.length; i++) {
-              imagesAddr.push(element.filename);
+              imagesAddr.push(res.req.files[i].filename);
             }
             await Pet.findOneAndUpdate(
               {
@@ -201,6 +238,16 @@ router
       console.log(error);
     }
   })
+  /**
+   * PATCH
+   * add or remove photos
+   */
+  //Todo
+  .patch((req, res) => {})
+  /**
+   * DELETE
+   * delete pet's photos
+   */
   .delete(async (req, res) => {
     try {
       if (req.user != null) {
@@ -259,6 +306,240 @@ router
         }
       } else {
         res.sendStatus(401);
+      }
+    } catch (error) {
+      console.log(error);
+    }
+  });
+
+router
+  .route("/products")
+  .post((req, res) => {
+    try {
+      if (req.user != null) {
+        if (req.user.admin) {
+          const upload = multer({ storage: storage }).array("images", 6);
+          upload(req, res, async (err) => {
+            if (err) {
+              console.log(err);
+            } else {
+              const imagesAddr = [];
+              for (let i = 0; i < res.req.files.length; i++) {
+                imagesAddr.push(res.req.files[i].filename);
+              }
+              await Product.findOneAndUpdate(
+                {
+                  _id: req.query.id,
+                },
+                {
+                  $push: {
+                    photos: {
+                      $each: imagesAddr,
+                    },
+                  },
+                }
+              ).then(() => {
+                res.status(200).send("images uploaded correctly");
+              });
+            }
+          });
+        } else {
+          const upload = multer({ storage: storage }).array("images", 6);
+          upload(req, res, async (err) => {
+            if (err) {
+              console.log(err);
+            } else {
+              const imagesAddr = [];
+              for (let i = 0; i < res.req.files.length; i++) {
+                imagesAddr.push(res.req.files[i].filename);
+              }
+              await Product.findOneAndUpdate(
+                {
+                  $and: [{ _id: req.query.id }, { seller: req.user.username }],
+                },
+                {
+                  $push: {
+                    pictures: {
+                      $each: imagesAddr,
+                    },
+                  },
+                }
+              ).then(() => {
+                res.status(200).send("images uploaded correctly");
+              });
+            }
+          });
+        }
+      } else {
+        res.sendStatus(401);
+      }
+    } catch (error) {
+      console.log(error);
+    }
+  })
+
+  .patch((req, res) => {
+    if (req.user != null) {
+      if (req.user.admin) {
+      }
+    }
+  });
+
+router
+  .route("/posts")
+  .post((req, res) => {
+    try {
+      if (req.user != null) {
+        if (req.user.admin) {
+          const upload = multer({ storage: storage }).array("images");
+          upload(req, res, async (err) => {
+            if (err) {
+              console.log(err);
+            } else {
+              const imagesAddr = [];
+              for (let i = 0; i < res.req.files.length; i++) {
+                imagesAddr.push(res.req.files[i].filename);
+              }
+              await Post.findOneAndUpdate(
+                {
+                  _id: req.query.id,
+                },
+                {
+                  $push: {
+                    photos: {
+                      $each: imagesAddr,
+                    },
+                  },
+                }
+              );
+              res.status(200).send("images uploaded correctly");
+            }
+          });
+        } else {
+          const upload = multer({ storage: storage }).array("images");
+          upload(req, res, async (err) => {
+            if (err) {
+              console.log(err);
+            } else {
+              const imagesAddr = [];
+              for (let i = 0; i < res.req.files.length; i++) {
+                imagesAddr.push(element.filename);
+              }
+              await Post.findOneAndUpdate(
+                {
+                  $and: [
+                    { _id: req.query.id },
+                    { username: req.user.username },
+                  ],
+                },
+                {
+                  $push: {
+                    photos: {
+                      $each: imagesAddr,
+                    },
+                  },
+                }
+              );
+              res.status(200).send("images uploaded correctly");
+            }
+          });
+        }
+      }
+    } catch (error) {
+      console.log(error);
+    }
+  })
+  .patch((req, res) => {});
+
+router
+  .route("/services")
+  /**
+   * TODO: get
+   */
+  .get((req, res) => {})
+  /**
+   * POST
+   * create new service with photos
+   */
+  .post((req, res) => {
+    try {
+      if (req.user != null) {
+        if (req.user.admin) {
+          const upload = multer({ storage: storage }).array("images", 2);
+          upload(req, res, async (err) => {
+            if (err) {
+              console.log(err);
+            } else {
+              const imagesAddr = [];
+              for (let i = 0; i < res.req.files.length; i++) {
+                imagesAddr.push(element.filename);
+              }
+              await Service.findOneAndUpdate(
+                {
+                  _id: req.query.id,
+                },
+                {
+                  $push: {
+                    pictures: {
+                      $each: imagesAddr,
+                    },
+                  },
+                }
+              );
+              res.status(200).send("images uploaded correctly");
+            }
+          });
+        } else {
+          res.sendStatus(401);
+        }
+      }
+    } catch (error) {
+      console.log(error);
+    }
+  })
+  /**
+   * PATCH
+   * edit service's photo
+   */
+  .patch();
+
+router
+  .route("/news")
+  /**
+   * POST
+   * add photos to news
+   */
+  .post(async (req, res) => {
+    try {
+      if (req.user != null) {
+        if (req.user.admin) {
+          const upload = multer({ storage: storage }).array("images", 2);
+          upload(req, res, async (err) => {
+            if (err) {
+              console.log(err);
+            } else {
+              const imagesAddr = [];
+              for (let i = 0; i < res.req.files.length; i++) {
+                imagesAddr.push(element.filename);
+              }
+              await News.findOneAndUpdate(
+                {
+                  _id: req.query.id,
+                },
+                {
+                  $push: {
+                    pictures: {
+                      $each: imagesAddr,
+                    },
+                  },
+                }
+              );
+              res.status(200).send("images uploaded correctly");
+            }
+          });
+        } else {
+          res.sendStatus(401);
+        }
       }
     } catch (error) {
       console.log(error);
